requirements.txt
anyio==4.0.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.3
colorama==0.4.6
blinker==1.7.0
Flask==2.2.5
Flask-Cors==4.0.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
polling2==0.5.0
python-engineio==4.8.0
python-socketio==5.10.0
setuptools==68.2.2
simple-websocket==1.0.0
sniffio==1.3.0
speechmatics-python==1.12.0
tenacity==8.2.3
toml==0.10.2
websockets==12.0
Werkzeug==2.2.3
wsproto==1.2.0